+++ src/components/Header.astro
@@ -1,6 +1,6 @@
 ---
 import { MobileMenu } from "@/components/MobileMenu";
-import { ThemeToggleAnimated } from "@/components/ThemeToggleAnimated";
+import { ThemeToggleAnimated } from "@/components/theme/ThemeToggleAnimated";
 
 const navItems = [
   { href: "/", label: "Home" },
