--- src/components/ThemeToggle.tsx
@@ -1,77 +1,0 @@
-import { useEffect, useState } from "react";
-import { Button } from "@/components/ui/button";
-
-// DOM状態から現在のテーマを取得する関数
-// ThemeInit.astroで設定されたクラスを読み取る
-function getInitialTheme(): "light" | "dark" {
-  if (typeof document !== "undefined") {
-    return document.documentElement.classList.contains("dark") ? "dark" : "light";
-  }
-  return "light";
-}
-
-export function ThemeToggle() {
-  // SSR時は固定値を使用し、ハイドレーションミスマッチを防止
-  const [theme, setTheme] = useState<"light" | "dark">("light");
-
-  // マウント後にDOM状態と同期（ThemeInit.astroで設定済みのクラスを読み取る）
-  useEffect(() => {
-    setTheme(getInitialTheme());
-  }, []);
-
-  const toggleTheme = () => {
-    const newTheme = theme === "light" ? "dark" : "light";
-    setTheme(newTheme);
-    localStorage.setItem("theme", newTheme);
-    document.documentElement.classList.toggle("dark", newTheme === "dark");
-  };
-
-  return (
-    <Button
-      variant="ghost"
-      size="icon"
-      onClick={toggleTheme}
-      aria-label={theme === "light" ? "ダークモードに切り替え" : "ライトモードに切り替え"}
-    >
-      {theme === "light" ? (
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="20"
-          height="20"
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          aria-hidden="true"
-        >
-          <path d="M12 3a6 6 0 0 0 9 9 9 9 0 1 1-9-9Z" />
-        </svg>
-      ) : (
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="20"
-          height="20"
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          aria-hidden="true"
-        >
-          <circle cx="12" cy="12" r="4" />
-          <path d="M12 2v2" />
-          <path d="M12 20v2" />
-          <path d="m4.93 4.93 1.41 1.41" />
-          <path d="m17.66 17.66 1.41 1.41" />
-          <path d="M2 12h2" />
-          <path d="M20 12h2" />
-          <path d="m6.34 17.66-1.41 1.41" />
-          <path d="m19.07 4.93-1.41 1.41" />
-        </svg>
-      )}
-    </Button>
-  );
-}
