--- src/components/icons/index.ts
@@ -1,2 +1,0 @@
-export { MoonIcon } from "./MoonIcon";
-export { SunIcon } from "./SunIcon";
