+++ src/components/scrapbox/QueryProvider.tsx
@@ -1,5 +1,5 @@
 import { QueryClientProvider } from "@tanstack/react-query";
-import { queryClient } from "@/lib/queryClient";
+import { queryClient } from "./queryClient";
 
 interface Props {
   children: React.ReactNode;
