+++ src/components/scrapbox/ScrapboxCard.test.tsx
@@ -1,7 +1,7 @@
 import { act, render, screen } from "@testing-library/react";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import type { ScrapboxPageData } from "@/types/scrapbox";
 import { ScrapboxCard } from "./ScrapboxCard";
+import type { ScrapboxPageData } from "./types";
 
 class MockIntersectionObserver {
   private callback: IntersectionObserverCallback;
