+++ src/components/scrapbox/useScrapboxData.test.tsx
@@ -2,7 +2,7 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { renderHook, waitFor } from "@testing-library/react";
 import type { ReactNode } from "react";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import type { ScrapboxPageData } from "@/types/scrapbox";
+import type { ScrapboxPageData } from "./types";
 import { useScrapboxData } from "./useScrapboxData";
 
 function createWrapper() {
