+++ src/components/search/CommandPalette.test.tsx
