+++ src/layouts/BaseHead.astro
@@ -1,5 +1,5 @@
 ---
-import ThemeInit from "@/components/ThemeInit.astro";
+import ThemeInit from "@/components/theme/ThemeInit.astro";
 import { buildPageStructuredData, serializeJsonLd } from "@/lib/structuredData";
 
 interface Props {
