+++ src/layouts/BaseLayout.astro
@@ -1,8 +1,8 @@
 ---
 import "../styles/globals.css";
-import { CommandPalette } from "@/components/CommandPalette";
 import Footer from "@/components/Footer.astro";
 import Header from "@/components/Header.astro";
+import { CommandPalette } from "@/components/search/CommandPalette";
 import BaseHead from "./BaseHead.astro";
 
 interface Props {
