+++ src/pages/search.astro
@@ -1,7 +1,7 @@
 ---
 import { getCollection } from "astro:content";
 import PageHero from "@/components/PageHero.astro";
-import { SearchBox } from "@/components/SearchBox";
+import { SearchBox } from "@/components/search/SearchBox";
 import { TagBubbleCloud } from "@/components/TagBubbleCloud";
 import BaseLayout from "@/layouts/BaseLayout.astro";
